store: Initial support for purchases with 0 price
