+++ Instanssi/store/static/store/css/style.scss
@@ -16,7 +16,6 @@
     z-index: 1;
     background: #fff;
 
-    box-shadow: rgba(#000, .2) 0 2px 8px;
 
     @media screen and (max-width: $screen-xs-max) {
         // Fill screen width on smaller screens.
@@ -25,6 +24,8 @@
         padding-left: $spacer * 2;
         padding-right: $spacer * 2;
         max-width: 100vw;
+
+        box-shadow: rgba(#000, .2) 0 2px 8px;
     }
 
     ul.nav {
